🔀 Restructure: jwt 관련 파일 경로 수정
